// 12_and_12.1_home_work/src/types/interfaces.ts
export interface ISource{
  id: string | null;
  name: string;
}

export interface IArticle {
  source: ISource;  
  author: string | null;
  title: string;
  description: string | null;
  url: string;
  urlToImage: string | null;
  publishedAt: string;
  content: string | null;
}

export interface INewsResponse{
  status: string;
  totalResults: number;
  articles: IArticle[]
}

export  interface ISourseResponse{
  status: string;
  source: ISource[];
}